refactor(protocol): modularize protocol and fsm monoliths
- Dismantled `protocol.py` and `protocol_fsm.py` monoliths into a new `src/ramses_tx/protocol/` package.
- Extracted foundational layers into `base.py` and discrete states into `fsm.py`.
- Integrated elevated system sync collision avoidance natively into the FSM queue processing.
- Extracted concrete implementations into `core.py` and assemblers into `factory.py`.
- Updated test suite imports to point to the new package structure.
- Maintained 100% Mypy strict compliance and Ruff linting standards.
