+++ tests/tests_rf/test_use_regex.py
@@ -58,7 +58,9 @@ TESTS_INBOUND = {  # sent by other, received
 
 @pytest.fixture(autouse=True)
 def patches_for_tests(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setattr("ramses_tx.protocol._DBG_DISABLE_IMPERSONATION_ALERTS", True)
+    monkeypatch.setattr(
+        "ramses_tx.protocol.core._DBG_DISABLE_IMPERSONATION_ALERTS", True
+    )
     monkeypatch.setattr("ramses_tx.transport.port.MIN_INTER_WRITE_GAP", 0)
